pyproject.toml
[tool.poetry]
name = "pagesagent"
version = "0.1.0"
description = ""
authors = ["Frank Lin"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
notion-client = "^2.2.1"
markdown = "^3.6"
bs4 = "^0.0.2"
networkx = "^3.3"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
notional = "^0.8.2"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
